MIA-412 staff search (#706)
* MIA-412 staff search

+++ src/main/java/uk/gov/justice/digital/hmpps/keyworker/integration/nomisuserroles/JobClassification.kt
@@ -1,12 +1,13 @@
 package uk.gov.justice.digital.hmpps.keyworker.integration.nomisuserroles
 
 import com.fasterxml.jackson.annotation.JsonAlias
+import java.math.BigDecimal
 import java.time.LocalDate
 
 interface JobClassification {
   val position: String
   val scheduleType: String
-  val hoursPerWeek: Int
+  val hoursPerWeek: BigDecimal
   val fromDate: LocalDate
   val toDate: LocalDate?
 }
@@ -14,7 +15,7 @@ interface JobClassification {
 data class StaffJobClassificationRequest(
   override val position: String,
   override val scheduleType: String,
-  override val hoursPerWeek: Int,
+  override val hoursPerWeek: BigDecimal,
   override val fromDate: LocalDate,
   override val toDate: LocalDate?,
 ) : JobClassification
@@ -25,7 +26,7 @@ data class StaffJobClassification(
   val staffId: Long,
   override val position: String,
   override val scheduleType: String,
-  override val hoursPerWeek: Int,
+  override val hoursPerWeek: BigDecimal,
   override val fromDate: LocalDate,
   override val toDate: LocalDate?,
 ) : JobClassification {
